+++ sound.h
@@ -3,7 +3,7 @@
 */
 
 void SOUND_init();
-void SOUND_beep(int freq, int length);
+void SOUND_beep(double freq, int length);
 void soundfail();
 void soundunlock();
 void soundfailR();
